Moved HotelSortArgument from Request class to its own file.
