+++ HotelReservation/src/main/java/com/subham/hotelreservation/models/Request.java
@@ -22,21 +22,4 @@ public class Request {
     public CustomerType getType(){
         return type;
     }
-
-    public static class HotelSortArgument {
-        private Double amount;
-        private Hotel h;
-
-        public HotelSortArgument(double amount, Hotel h) {
-            this.amount = amount;
-            this.h = h;
-        }
-
-        public Hotel getHotel() {
-            return h;
-        }
-        public double getAmount(){
-            return amount;
-        }
-    }
 }
