+++ HotelReservation/src/main/java/com/subham/hotelreservation/utils/Comp.java
@@ -1,12 +1,13 @@
 package com.subham.hotelreservation.utils;
 
+import com.subham.hotelreservation.models.HotelSortArgument;
 import com.subham.hotelreservation.models.Request;
 
 import java.util.Comparator;
 
-public class Comp implements Comparator<Request.HotelSortArgument> {
+public class Comp implements Comparator<HotelSortArgument> {
 
-    public int compare(Request.HotelSortArgument hotelSortArgument, Request.HotelSortArgument t1) {
+    public int compare(HotelSortArgument hotelSortArgument, HotelSortArgument t1) {
         if(hotelSortArgument.getAmount() != t1.getAmount()){
             return (int)(hotelSortArgument.getAmount() - t1.getAmount());
         }
